+++ uefi-test-runner/src/fs/mod.rs
@@ -2,9 +2,9 @@
 
 use alloc::string::{String, ToString};
 use alloc::vec::Vec;
+use uefi::boot::ScopedProtocol;
 use uefi::fs::{FileSystem, IoError, IoErrorContext, PathBuf};
 use uefi::proto::media::fs::SimpleFileSystem;
-use uefi::table::boot::ScopedProtocol;
 use uefi::{cstr16, fs, Status};
 
 /// Tests functionality from the `uefi::fs` module. This test relies on a